feat(common): LOCAL-3191 rename some variables
