+++ apps/storefront/src/components/layout/B3MainHeader.tsx
@@ -14,7 +14,7 @@ import { getContrastColor } from '../outSideComponents/utils/b3CustomStyles';
 
 import B3AccountInfo from './B3AccountInfo';
 import B3CompanyHierarchy from './B3CompanyHierarchy';
-import Locales from './B3LocaleDropdown';
+import Locales from './Locales';
 import B3StatusNotification from './B3StatusNotification';
 
 export default function MainHeader({ title }: { title: string }) {
